Fix scan unknown marker inheritance.
