+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri/src/org/eclipse/chemclipse/chromatogram/xxd/calculator/supplier/amdiscalri/settings/AbstractAlkaneSettings.java
@@ -1,0 +1,62 @@
+/*******************************************************************************
+ * Copyright (c) 2024 Lablicate GmbH.
+ * 
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ * 
+ * Contributors:
+ * Philip Wenig - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.settings;
+
+import org.eclipse.chemclipse.model.identifier.AbstractIdentifierSettings;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class AbstractAlkaneSettings extends AbstractIdentifierSettings implements IAlkaneSettings {
+
+	@JsonProperty(value = "Number of Targets", defaultValue = "15")
+	private int numberOfTargets = 15;
+	@JsonProperty(value = "Min Match Factor", defaultValue = "80.0")
+	private float minMatchFactor = 80.0f;
+	@JsonProperty(value = "Min Reverse Match Factor", defaultValue = "80.0")
+	private float minReverseMatchFactor = 80.0f;
+
+	@Override
+	public int getNumberOfTargets() {
+
+		return numberOfTargets;
+	}
+
+	@Override
+	public void setNumberOfTargets(int numberOfTargets) {
+
+		this.numberOfTargets = numberOfTargets;
+	}
+
+	@Override
+	public float getMinMatchFactor() {
+
+		return minMatchFactor;
+	}
+
+	@Override
+	public void setMinMatchFactor(float minMatchFactor) {
+
+		this.minMatchFactor = minMatchFactor;
+	}
+
+	@Override
+	public float getMinReverseMatchFactor() {
+
+		return minReverseMatchFactor;
+	}
+
+	@Override
+	public void setMinReverseMatchFactor(float minReverseMatchFactor) {
+
+		this.minReverseMatchFactor = minReverseMatchFactor;
+	}
+}
