+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri/src/org/eclipse/chemclipse/chromatogram/xxd/calculator/supplier/amdiscalri/settings/IAlkaneSettings.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2024 Lablicate GmbH.
- *
+ * Copyright (c) 2024 Lablicate GmbH.
+ * 
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -11,7 +11,17 @@
  *******************************************************************************/
 package org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.settings;
 
-import org.eclipse.chemclipse.chromatogram.msd.identifier.settings.IMassSpectrumIdentifierSettings;
+public interface IAlkaneSettings {
 
-public class MassSpectrumIdentifierAlkaneSettings extends AbstractAlkaneSettings implements IMassSpectrumIdentifierSettings {
+	int getNumberOfTargets();
+
+	void setNumberOfTargets(int numberOfTargets);
+
+	float getMinMatchFactor();
+
+	void setMinMatchFactor(float minMatchFactor);
+
+	float getMinReverseMatchFactor();
+
+	void setMinReverseMatchFactor(float minReverseMatchFactor);
 }
