+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file/src/org/eclipse/chemclipse/chromatogram/xxd/identifier/supplier/file/settings/MassSpectrumUnknownSettings.java
@@ -11,20 +11,17 @@
  *******************************************************************************/
 package org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file.settings;
 
+import org.eclipse.chemclipse.chromatogram.msd.identifier.settings.IMassSpectrumIdentifierSettings;
 import org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file.preferences.PreferenceSupplier;
-import org.eclipse.chemclipse.model.identifier.IIdentifierSettings;
+import org.eclipse.chemclipse.model.identifier.AbstractIdentifierSettings;
 import org.eclipse.chemclipse.support.settings.FloatSettingsProperty;
 import org.eclipse.chemclipse.support.settings.IntSettingsProperty;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 
-public class MassSpectrumUnknownSettings extends MassSpectrumIdentifierSettings implements IUnknownSettingsMSD {
+public class MassSpectrumUnknownSettings extends AbstractIdentifierSettings implements IUnknownSettingsMSD, IMassSpectrumIdentifierSettings {
 
-	@JsonProperty(value = "Limit Match Factor", defaultValue = "80.0")
-	@JsonPropertyDescription(value = "Run an identification if no target exists with a Match Factor >= the given limit.")
-	@FloatSettingsProperty(minValue = IIdentifierSettings.MIN_LIMIT_MATCH_FACTOR, maxValue = IIdentifierSettings.MAX_LIMIT_MATCH_FACTOR)
-	private float limitMatchFactor = 80.0f;
 	@JsonProperty(value = "Target Name", defaultValue = PreferenceSupplier.DEF_TARGET_NAME_UNKNOWN)
 	private String targetName = "Unknown";
 	@JsonProperty(value = "Match Quality", defaultValue = "80.0")
@@ -49,18 +46,6 @@ public class MassSpectrumUnknownSettings extends MassSpectrumIdentifierSettings 
 	private boolean includeRetentionIndex = false;
 
 	@Override
-	public float getLimitMatchFactor() {
-
-		return limitMatchFactor;
-	}
-
-	@Override
-	public void setLimitMatchFactor(float limitMatchFactor) {
-
-		this.limitMatchFactor = limitMatchFactor;
-	}
-
-	@Override
 	public String getTargetName() {
 
 		return targetName;
@@ -82,11 +67,6 @@ public class MassSpectrumUnknownSettings extends MassSpectrumIdentifierSettings 
 	public void setMatchQuality(float matchQuality) {
 
 		this.matchQuality = matchQuality;
-	}
-
-	public int getNumberOfMZ() {
-
-		return numberOfMZ;
 	}
 
 	@Override
